+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/ProcessOutVariables.java
@@ -1,0 +1,48 @@
+package com.walmartlabs.concord.server.process;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2018 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.walmartlabs.concord.project.InternalConstants;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Collections;
+import java.util.Map;
+
+public final class ProcessOutVariables {
+
+    private static final ObjectMapper objectMapper = new ObjectMapper();
+
+    @SuppressWarnings("unchecked")
+    public static Map<String, Object> read(Path attachmetsDir) throws IOException {
+        Path processOut = attachmetsDir.resolve(InternalConstants.Files.OUT_VALUES_FILE_NAME);
+        if (!Files.exists(processOut)) {
+            return Collections.emptyMap();
+        }
+
+        return objectMapper.readValue(processOut.toFile(), Map.class);
+    }
+
+    private ProcessOutVariables() {
+    }
+}